backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
psycopg[binary]>=3.1.18

# Phase 3+
ib_insync==0.9.86
pandas>=2.2.3
numpy>=2.1.0

# Phase 5+
resend==0.7.0
apscheduler>=3.10
